feat: changed the data type of category_id in books to BIGINT

+++ database/sql_migrations/1_initiate.sql
@@ -19,7 +19,7 @@ CREATE TABLE books (
     price int NOT NULL,
     total_page int NOT NULL,
     thickness varchar(128) NOT NULL,
-    category_id int NOT NULL,
+    category_id BIGINT NOT NULL,
     created_at timestamp NOT NULL,
     created_by varchar(255) NOT NULL,
     modified_at timestamp NOT NULL,
